app/pyproject.toml
[project]
name = "learn-strands-agents"
version = "0.1.0"
description = "learn strands agents"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.14.3",
    "httpx>=0.28.1",
    "ruff>=0.14.13",
    "strands-agents>=1.22.0",
    "ty>=0.0.12",
]
